Add ability to create Snapshot recordings (#67)
* Extract custom recordings form component

* Add tabset around card body

* Add Snapshot recording form
